requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
watchdog>=4.0.0
schedule>=1.2.0
anthropic>=0.34.0
fpdf2>=2.7.9
